Update content gateway according to push items     - Added mechanism to add, update or delete the Content Gateway (CGW) metadata through ymal metadata     - Added ymal schema validation     - Updated entry points to read the user credentials from CLI for the CGW operations
